requirements.txt
alembic==1.13.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
fsspec==2023.12.2
greenlet==3.0.3
huggingface-hub==0.20.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
Mako==1.3.0
MarkupSafe==2.1.3
mpmath==1.2.1
networkx==3.0rc1
numpy==1.26.2
packaging==23.2
pdf2image==1.16.3
Pillow==10.1.0
pytesseract==0.3.10
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
SQLAlchemy==2.0.23
sympy==1.11.1
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.3.0.dev20231222+cpu
torchaudio==2.2.0.dev20231222+cpu
torchvision==0.18.0.dev20231222+cpu
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
